Save the command line along with the state file
We need to save the command line arguments passed to qemu when saving
the state file. This is because qemu expects the same arguments when
restoring from the saved state and it does not store it in the state
file. Now we read the cmd file at the time of state restoration.