Remove erraneous loop in qsim loader
The run loop returns once loading is done and the application is about to start
executing. We need to swap the context and set any callbacks needed by the
benchmark. The loop here does not allow that. It will possibly execute the
benchmark without the callbacks being set. This commit fixes the issue.